Not load last message contacts when chat window will show
This created a race condition since the session could be in the model or not,
all happens on the same event. If ithe session was not in the model, stream add
did fail. The race condition could occur when accepting audio/video/chat
sessions and the window show was called.